+++ src/main/java/me/drex/essentials/command/impl/home/SetHomeCommand.java
@@ -58,7 +58,7 @@ public class SetHomeCommand extends Command {
             src.sendFailure(Component.nullToEmpty("您無法在此維度建立 Home點"));
             return FAILURE;
         }
-        PlayerData playerData = DataStorage.getOfflinePlayerData(src.getServer(), target.getId());
+        PlayerData playerData = DataStorage.getOfflinePlayerData(src.getServer(), target);
         Map<String, Home> homes = playerData.homes;
         Home previousHome = homes.get(name);
         if (previousHome != null && !confirm) {
@@ -77,7 +77,7 @@ public class SetHomeCommand extends Command {
         }
         Home home = new Home(new Location(src));
         homes.put(name, home);
-        DataStorage.updateOfflinePlayerData(src.getServer(), target.getId(), playerData);
+        DataStorage.updateOfflinePlayerData(src.getServer(), target, playerData);
         if (self) {
             src.sendSystemMessage(localized("fabric-essentials.commands.sethome.self", home.placeholders(name)));
         } else {